# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Azora OS API
  description: |
    Constitutional AI Operating System API
    
    **Ubuntu Principle:** "My API strengthens our foundation"
    
    Complete API documentation for all Azora OS services.
  version: 3.0.0
  contact:
    name: Azora OS Team
    url: https://azora.world
    email: api@azora.world
  license:
    name: Proprietary
    url: https://azora.world/license
servers:
  - url: http://localhost:4000/api
    description: Local development
  - url: https://api.azora.world
    description: Production
...
paths:
...
  /courses:
    get:
      tags: [Education]
      summary: List all courses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                type: object
                properties:
                  courses:
                    type: array
                    items:
                      $ref: '#/components/schemas/Course'
                  total:
                    type: integer
                  page:
                    type: integer
                  pages:
                    type: integer

    post:
      tags: [Education]
      summary: Create new course
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseInput'
      responses:
        '201':
          description: Course created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        instructorId:
          type: string
        price:
          type: number
        currency:
          type: string
        duration:
          type: integer
        level:
          type: string
        category:
          type: string
        published:
          type: boolean
    CourseInput:
      type: object
      required: [title, description, duration, level, category]
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
          default: 0
        duration:
          type: integer
        level:
          type: string
        category:
          type: string
...